requirements.txt
beautifulsoup4>=4.12.0
html5lib>=1.1
lxml>=4.9.0
pydantic>=2.0.0
openai>=1.0.0
anthropic>=0.18.0
python-dotenv>=1.0.0
